+++ Events.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
@@ -16,23 +17,23 @@ use humhub\modules\legal\widgets\Content;
 use humhub\modules\legal\widgets\CookieNote;
 use humhub\modules\post\models\Post;
 use humhub\modules\rest\components\BaseController;
+use humhub\modules\ui\menu\MenuLink;
 use humhub\modules\user\models\forms\Registration;
 use humhub\modules\user\models\User;
+use humhub\modules\user\widgets\AccountSettingsMenu;
 use humhub\widgets\LayoutAddons;
 use Yii;
 use yii\base\ActionEvent;
 use yii\helpers\Url;
 use yii\web\UserEvent;
 
-
 /**
  * @author luke
  */
 class Events
 {
-
-    const SESSION_KEY_LEGAL_CHECK = 'legalModuleChecked';
-    const SESSION_KEY_LEGAL_AFTER_REGISTRATION = 'legalModuleAfterRegistration';
+    public const SESSION_KEY_LEGAL_CHECK = 'legalModuleChecked';
+    public const SESSION_KEY_LEGAL_AFTER_REGISTRATION = 'legalModuleAfterRegistration';
 
     public static function onFooterMenuInit($event)
     {
@@ -49,11 +50,11 @@ class Events
             $page = Page::getPage($pageKey);
             if ($page !== null) {
                 $sortOrder += 10;
-                $event->sender->addItem(array(
+                $event->sender->addItem([
                     'label' => $page->title,
                     'url' => Url::to(['/legal/page/view', 'pageKey' => $pageKey], true),
                     'sortOrder' => $sortOrder,
-                ));
+                ]);
             }
         }
 
@@ -141,7 +142,7 @@ class Events
             $event->result = Yii::$app->response->redirect(['/legal/page/update']);
         }
         // Show legal pages in full screen with confirm form, one by one (after account creation)
-        elseif($registrationCheck->showTermsCheck() || $registrationCheck->showPrivacyCheck()) {
+        elseif ($registrationCheck->showTermsCheck() || $registrationCheck->showPrivacyCheck()) {
             $event->isValid = false;
             $event->result = Yii::$app->response->redirect(['/legal/page/confirm']);
         }
@@ -180,27 +181,12 @@ class Events
         /** @var Module $module */
         $module = Yii::$app->getModule('legal');
 
-        $hForm->models['RegistrationChecks'] = new RegistrationChecks(['restrictToSettingKey' => $module->showPagesAfterRegistration() ? RegistrationChecks::SETTING_KEY_AGE : false]);
+        $model = new RegistrationChecks(['restrictToSettingKey' => $module->showPagesAfterRegistration() ? RegistrationChecks::SETTING_KEY_AGE : false]);
+        $hForm->models['RegistrationChecks'] = $model;
 
         if ($module->showPagesAfterRegistration()) {
             Yii::$app->session->set(static::SESSION_KEY_LEGAL_AFTER_REGISTRATION, 'true');
         }
-    }
-
-    public static function onRegistrationFormRender($event)
-    {
-        if (static::skipVerifying()) {
-            return;
-        }
-
-        /** @var Registration $hForm */
-        $hForm = $event->sender;
-
-        /** @var RegistrationChecks $model */
-        $model = $hForm->models['RegistrationChecks'];
-
-        /** @var Module $module */
-        $module = Yii::$app->getModule('legal');
 
         $elements = [];
 
@@ -210,6 +196,7 @@ class Events
                 'class' => 'form-control',
             ];
         }
+
         if ($model->showPrivacyCheck()) {
             $elements['dataPrivacyCheck'] = [
                 'type' => 'checkbox',
@@ -226,7 +213,7 @@ class Events
 
         $hForm->definition['elements']['RegistrationChecks'] = [
             'type' => 'form',
-            'elements' => $elements
+            'elements' => $elements,
         ];
     }
 
@@ -263,5 +250,31 @@ class Events
         if ($richText->record instanceof Post || $richText->record instanceof Comment) {
             $event->result = Content::widget(['content' => $event->result, 'richtext' => false]);
         }
+    }
+
+    public static function onAccountSettingsMenuInit($event)
+    {
+        /* @var AccountSettingsMenu $menu */
+        $menu = $event->sender;
+
+        /* @var Module $module */
+        $module = Yii::$app->getModule('legal');
+
+        if ($module->isEnabledExportUserData()) {
+            $menu->addEntry(new MenuLink([
+                'label' => Yii::t('LegalModule.base', 'Export personal data'),
+                'url' => ['/legal/export'],
+                'sortOrder' => 1000,
+                'isActive' => MenuLink::isActiveState('legal', 'export'),
+            ]));
+        }
+    }
+
+    /**
+     * Callback on daily cron job run
+     */
+    public static function onCronDailyRun()
+    {
+        Yii::$app->queue->push(new jobs\DeletePackages());
     }
 }