+++ Module.php
@@ -13,6 +13,10 @@ use yii\helpers\Url;
 
 class Module extends \humhub\components\Module
 {
+    /**
+     * @inheritdoc
+     */
+    public $resourcesPath = 'resources';
 
     /**
      * @inheritdoc
@@ -52,7 +56,7 @@ class Module extends \humhub\components\Module
      */
     public function showPagesAfterRegistration()
     {
-        return (boolean)$this->settings->get('showPagesAfterRegistration', false);
+        return (bool)$this->settings->get('showPagesAfterRegistration', false);
     }
 
     /**
@@ -80,12 +84,24 @@ class Module extends \humhub\components\Module
         return $this->settings->get('defaultLanguage');
     }
 
-    /**
-     * @return bool
-     */
-    public function showAgeCheck()
+    public function isAllowedExportUserData(): bool
     {
-        return (boolean)$this->settings->get('showAgeCheck', false);
+        return Yii::$app->hasModule('rest') && Yii::$app->getModule('rest')->isActivated;
+    }
+
+    public function isEnabledExportUserData(): bool
+    {
+        return $this->isAllowedExportUserData() && $this->settings->get('exportUserData', false);
+    }
+
+    public function getExportUserDays(): int
+    {
+        return (int) $this->settings->get('exportUserDays', 1);
+    }
+
+    public function showAgeCheck(): bool
+    {
+        return (bool) $this->settings->get('showAgeCheck', false);
     }
 
     /**
@@ -94,16 +110,6 @@ class Module extends \humhub\components\Module
     public function getMinimumAge()
     {
         return $this->settings->get('minimumAge', 16);
-    }
-
-    public function getName()
-    {
-        return Yii::t('LegalModule.base', 'Legal Tools');
-    }
-
-    public function getDescription()
-    {
-        return Yii::t('LegalModule.base', 'Adds several editable legal options, like an imprint and a privacy policy.');
     }
 
 }