+++ assets/Assets.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
@@ -11,19 +12,18 @@ use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
-
     public $defer = true;
 
     public $publishOptions = [
-        'forceCopy' => false
+        'forceCopy' => false,
     ];
 
     public $sourcePath = '@legal/resources';
     public $css = [
-        'cookieconsent.min.css'
+        'cookieconsent.min.css',
     ];
     public $js = [
-        'cookieconsent.min.js'
+        'cookieconsent.min.js',
     ];
 
 }