+++ assets/ContentAssets.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2022 HumHub GmbH & Co. KG
@@ -23,7 +24,7 @@ class ContentAssets extends AssetBundle
      * @inheritdoc
      */
     public $js = [
-        'js/humhub.legal.js'
+        'js/humhub.legal.js',
     ];
 
     /**