+++ tests/config/test.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2022 HumHub GmbH & Co. KG
@@ -7,5 +8,5 @@
 
 return [
     'modules' => ['legal'],
-    'fixtures' => ['default']
+    'fixtures' => ['default'],
 ];